Consolidate test mtbl writing.
